Rename createJwt.js utility file to jwt.js
