+++ app/server/app/utilities/jwt.js
